requirements.txt
beautifulsoup4
Flask
pyTelegramBotAPI
matplotlib
numpy
dateparser
python-dotenv
requests
